Enable customized concurrency
Provide a way for user to customize concurrency of tasks to be executed.

This is extremely relevant when you try to use multirun to execute
a larget number of targets on a powerful machine, each affecting a
weaker downstream machine. E.g. using this to push multiple container
images could easily overwhelm the container registry, or reach a
ratelimit enforced by a cloud vendor.

Replace boolean 'parallel' attribute with int 'jobs' attribute which
determine the maximum concurrency of multirun. 'jobs' default value is
set to 1 which is the same with sequential execution.